+++ package.json
@@ -20,7 +20,8 @@
     "redux": "^4.0.5",
     "redux-devtools-extension": "^2.13.8",
     "redux-thunk": "^2.3.0",
-    "web-vitals": "^1.1.0"
+    "web-vitals": "^1.1.0",
+    "whatwg-fetch": "^3.6.2"
   },
   "scripts": {
     "start": "react-scripts start",
