+++ src/__tests__/Coin.test.js
@@ -4,19 +4,13 @@ import { getCoin } from '../actions/actionCreator';
 import { server, rest } from '../testServer';
 
 it('fetchs data correctly', async () => {
-  const response = await getCoin('BTC','USD');
-  expect(response).not.toBeNull;
+  const response = await getCoin('BTC', 'USD');
   expect(response.length).toEqual(1);
 });
 
 it('handles failure', async () => {
   server.use(
-    rest.get(
-      'https://api.nomics.com/v1/currencies/ticker',
-      (_req, res, ctx) => {
-        return res(ctx.status(404));
-      },
-    ),
+    rest.get('https://api.nomics.com/v1/currencies/ticker', (_req, res, ctx) => res(ctx.status(404))),
   );
 
   await expect(getCoin('BTC', 'USD')).rejects.toBeCalled;
