+++ src/__tests__/CoinList.test.js
@@ -5,7 +5,6 @@ import { server, rest } from '../testServer';
 
 it('fetchs data correctly', async () => {
   const response = await getCoinList('USD', 1);
-  expect(response).not.toBeNull;
   expect(response.length).toEqual(1);
 });
 
@@ -13,9 +12,7 @@ it('handles failure', async () => {
   server.use(
     rest.get(
       'https://api.nomics.com/v1/currencies/ticker',
-      (_req, res, ctx) => {
-        return res(ctx.status(404));
-      },
+      (_req, res, ctx) => res(ctx.status(404)),
     ),
   );
